* Fix the expand URLs * Up the max filesize to 25m * Upgrade Kotlin and add mysql-socket-factory * Redistribute the Database files and add JDBCDatabase * Add an option to print the config out * Add an option for the max workerExecuteTime
